Improvements in ProtocolLoop. Changed polling to be once per loop for both reads and writes. Added a max processing time per frame. Randomized peer order for read/write to avoid structural bias.

+++ src/hornetlib/net/receive.h
@@ -29,7 +29,7 @@ std::unique_ptr<T> ReceiveMessage(const Socket& sock, protocol::Magic magic) {
     throw std::runtime_error("Received incorrect magic bytes.");
 
   const auto& factory = protocol::MessageFactory::Default();
-  return protocol::ParseMessage<T>(factory, magic, {buf.data(), *length});
+  return protocol::ParseMessage<T>(factory, magic, {buf.data(), static_cast<size_t>(*length)});
 }
 
 }  // namespace hornet::net
